[fix] Use torch_dtype when loading transformer model

+++ src/cogkit/utils/load.py
@@ -12,7 +12,9 @@ def load_pipeline(
     pipeline = DiffusionPipeline.from_pretrained(model_id_or_path, torch_dtype=dtype)
     if transformer_path is not None:
         pipeline.transformer.save_config(transformer_path)
-        pipeline.transformer = pipeline.transformer.from_pretrained(transformer_path)
+        pipeline.transformer = pipeline.transformer.from_pretrained(
+            transformer_path, torch_dtype=dtype
+        )
     if lora_model_id_or_path is not None:
         load_lora_checkpoint(pipeline, lora_model_id_or_path)
     return pipeline
